blobstore,lvol: support cluster size above 2GiB
Any occurrence of cluster size in blobstore was change from uint32_t
to uint64_t, in order to enable cluster sizes above 2GiB.

+++ lib/bdev/lvol/vbdev_lvol.c
@@ -77,7 +77,7 @@ end:
 }
 
 int
-vbdev_lvs_create(struct spdk_bdev *base_bdev, uint32_t cluster_sz,
+vbdev_lvs_create(struct spdk_bdev *base_bdev, uint64_t cluster_sz,
 		 spdk_lvs_op_with_handle_complete cb_fn, void *cb_arg)
 {
 	struct spdk_bs_dev *bs_dev;
